Moved all event classes into their own folder Optical lenses now can be washed to remove tint
